pyproject.toml
[project]
name = "sample-server"
version = "0.1.0"
description = "FastMCP server demonstrating OAuth Proxy pattern with Keycloak authentication"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
]
